+++ exo-go/src/application/suite_test.go
@@ -3,7 +3,7 @@ package application_test
 import (
 	"testing"
 
-	"github.com/Originate/exosphere/exo-go/src/util"
+	"github.com/Originate/exosphere/exo-go/src/osplus"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 )
@@ -17,7 +17,7 @@ var homeDir string
 
 var _ = BeforeSuite(func() {
 	var err error
-	homeDir, err = util.GetHomeDirectory()
+	homeDir, err = osplus.GetHomeDirectory()
 	if err != nil {
 		panic(err)
 	}
