+++ exo-go/src/cmd/add.go
@@ -6,9 +6,9 @@ import (
 	"path"
 
 	"github.com/Originate/exosphere/exo-go/src/config"
+	"github.com/Originate/exosphere/exo-go/src/osplus"
 	"github.com/Originate/exosphere/exo-go/src/template"
 	"github.com/Originate/exosphere/exo-go/src/types"
-	"github.com/Originate/exosphere/exo-go/src/util"
 	prompt "github.com/segmentio/go-prompt"
 	"github.com/spf13/cobra"
 )
@@ -42,7 +42,7 @@ var addCmd = &cobra.Command{
 		if err != nil {
 			panic(err)
 		}
-		subdirectories, err := util.GetSubdirectories(serviceTmpDir)
+		subdirectories, err := osplus.GetSubdirectories(serviceTmpDir)
 		if err != nil {
 			panic(err)
 		}
@@ -56,11 +56,11 @@ var addCmd = &cobra.Command{
 			fmt.Println(err)
 			os.Exit(1)
 		}
-		err = util.MoveDirectory(path.Join(serviceTmpDir, serviceRole), path.Join(appDir, serviceRole))
+		err = osplus.MoveDirectory(path.Join(serviceTmpDir, serviceRole), path.Join(appDir, serviceRole))
 		if err != nil {
 			panic(err)
 		}
-		if !util.DoesFileExist(path.Join(appDir, serviceRole, "service.yml")) {
+		if !osplus.DoesFileExist(path.Join(appDir, serviceRole, "service.yml")) {
 			var templateDir string
 			templateDir, err = template.CreateServiceTemplateDir(serviceRole)
 			if err != nil {
