+++ exo-go/src/cmd/deploy.go
@@ -8,8 +8,8 @@ import (
 	"github.com/Originate/exosphere/exo-go/src/config"
 	"github.com/Originate/exosphere/exo-go/src/deploy"
 	"github.com/Originate/exosphere/exo-go/src/logger"
+	"github.com/Originate/exosphere/exo-go/src/osplus"
 	"github.com/Originate/exosphere/exo-go/src/types"
-	"github.com/Originate/exosphere/exo-go/src/util"
 	"github.com/spf13/cobra"
 )
 
@@ -27,7 +27,7 @@ var deployCmd = &cobra.Command{
 		if err != nil {
 			panic(err)
 		}
-		homeDir, err := util.GetHomeDirectory()
+		homeDir, err := osplus.GetHomeDirectory()
 		if err != nil {
 			panic(err)
 		}
