+++ exo-go/src/osplus/does_directory_exist.go
@@ -1,0 +1,7 @@
+package osplus
+
+// DoesDirectoryExist returns true if the directory dirPath is an existing directory,
+// and false otherwise
+func DoesDirectoryExist(dirPath string) bool {
+	return isDirectory(dirPath)
+}
