+++ exo-go/src/osplus/does_file_exist.go
@@ -1,0 +1,9 @@
+package osplus
+
+import "os"
+
+// DoesFileExist returns true if the file filePath exists, and false otherwise
+func DoesFileExist(filePath string) bool {
+	_, err := os.Stat(filePath)
+	return err == nil
+}
