+++ exo-go/src/osplus/is_directory.go
@@ -1,0 +1,8 @@
+package osplus
+
+import "os"
+
+func isDirectory(dirPath string) bool {
+	fi, err := os.Stat(dirPath)
+	return err == nil && fi.IsDir()
+}
