+++ exo-go/src/osplus/is_empty_file.go
@@ -1,0 +1,12 @@
+package osplus
+
+import "os"
+
+// IsEmptyFile returns true if filePath is an empty file, and false otherwise
+func IsEmptyFile(filePath string) bool {
+	fi, err := os.Stat(filePath)
+	if err != nil {
+		return false
+	}
+	return fi.Size() == 0
+}
